% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms.R
\name{get_all_aphiaids_by_aphiaid}
\alias{get_all_aphiaids_by_aphiaid}
\title{Get all synonymized AphiaIDs for an AphiaID.}
\usage{
get_all_aphiaids_by_aphiaid(aphiaid)
}
\description{
Get all synonymized AphiaIDs for an AphiaID.
}
