% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif.R
\name{get_gbif_dist}
\alias{get_gbif_dist}
\title{Get GBIF gridded distribution.}
\usage{
get_gbif_dist(
  scientificname = NULL,
  aphiaid = NULL,
  taxonkey = NULL,
  zoom = 0,
  square_size = 16
)
}
\description{
Get GBIF gridded distribution.
}
