% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{get_obis_dist}
\alias{get_obis_dist}
\title{Get OBIS gridded distribution.}
\usage{
get_obis_dist(scientificname = NULL, aphiaid = NULL, taxonkey = NULL, res = 3)
}
\description{
Get OBIS gridded distribution.
}
